requirements.txt
pywebview>=5.0.0
pywin32>=306
pystray>=0.19.5
Pillow>=10.0.0
pytest>=7.0.0
